Add integration tests for ABOP CLI and utility functions
- Introduced a new test module in `tests.rs` for comprehensive integration testing of the ABOP CLI.
- Added tests for constants, library path validation, database operations, and command-line option overrides.
- Created utility functions in `utils.rs` for fetching audiobook counts and displaying scan results.
- Implemented pagination logic for displaying audiobooks and ensured proper error handling.
- Updated `library_state.rs` to improve progress reporting during scanning.
- Enhanced `progress_cache.rs` to include a dedicated cache for simple percentage text.
