+++ abop-cli/src/commands/mod.rs
@@ -1,0 +1,7 @@
+//! Command implementations for ABOP CLI
+//!
+//! This module aggregates all command implementations, providing a clean
+//! separation between command parsing (in cli.rs) and command execution.
+
+pub mod scan;
+pub mod db;
